Puts

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /*
  * main-Entry piont
- * Return 0 Alway (success)
+ * Return: 0 Alway (success)
  */
 int main(void)
 {
@@ -11,10 +11,10 @@ int main(void)
 	long int longType;
 	long long int longlongType;
 
-	printf("Size of a char: %zu byte(s)\n", sizeof(charType));
-	printf("Size of an int: %zu byte(s)\n", sizeof(intType));
-	printf("Size of a Long int: %zu byte(s)\n", sizeof(longType));
-	printf("Size of a long long int: %zu byte(s)\n", sizeof(longlongType));
-	printF("Size of a float: %zu byte(s)\n", sizeof(floatType));
+	printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(charType));
+	printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(intType));
+	printf("Size of a Long int: %zu byte(s)\n", (unsigned long)sizeof(longType));
+	printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(longlongType));
+	printF("Size of a float: %zu byte(s)\n", (unsigned long)sizeof(floatType));
 	return (0);
 }
